# feat: improve stdlib base64 performance by 98%+