--- a/builtins_benchmark_test.go
+++ b/builtins_benchmark_test.go
@@ -38,3 +38,11 @@ func Benchmark_Builtin_substr(b *testing.B) {
 func Benchmark_Builtin_reverse(b *testing.B) {
 	RunBenchmark(b, "reverse")
 }
+
+func Benchmark_Builtin_base64(b *testing.B) {
+	RunBenchmark(b, "base64")
+}
+
+func Benchmark_Builtin_base64_byte_array(b *testing.B) {
+	RunBenchmark(b, "base64_byte_array")
+}