--- a/testdata/builtinBase64.jsonnet
+++ b/testdata/builtinBase64.jsonnet
@@ -1,0 +1,1 @@
+std.base64("hello")