--- a/testdata/builtinBase64_invalid_byte_array.golden
+++ b/testdata/builtinBase64_invalid_byte_array.golden
@@ -1,0 +1,10 @@
+RUNTIME ERROR: base64 encountered a non-integer value in the array, got string
+-------------------------------------------------
+	testdata/builtinBase64_invalid_byte_array:1:1-23	builtin function <base64>
+
+std.base64([1, "foo"])
+
+-------------------------------------------------
+	During evaluation	
+
+