--- a/testdata/builtinBase64_non_string_non_array.golden
+++ b/testdata/builtinBase64_non_string_non_array.golden
@@ -1,0 +1,10 @@
+RUNTIME ERROR: base64 can only base64 encode strings / arrays of single bytes, got number
+-------------------------------------------------
+	testdata/builtinBase64_non_string_non_array:1:1-14	builtin function <base64>
+
+std.base64(1)
+
+-------------------------------------------------
+	During evaluation	
+
+