--- a/testdata/builtinBase64_string_high_codepoint.golden
+++ b/testdata/builtinBase64_string_high_codepoint.golden
@@ -1,0 +1,10 @@
+RUNTIME ERROR: base64 encountered invalid codepoint value in the array (must be 0 <= X <= 255), got 256
+-------------------------------------------------
+	testdata/builtinBase64_string_high_codepoint:1:1-17	builtin function <base64>
+
+std.base64("Ā")
+
+-------------------------------------------------
+	During evaluation	
+
+