… builtin-benchmarks/base64.jsonnet
… builtin-benchmarks/base64_byte_array.jsonnet
… testdata/builtinBase64.golden
… testdata/builtinBase64_byte_array.golden
… testdata/builtinBase64_byte_array.jsonnet
… testdata/builtinBase64_invalid_byte_array.jsonnet
… testdata/builtinBase64_invalid_byte_array1.jsonnet
… testdata/builtinBase64_invalid_byte_array2.jsonnet
… testdata/builtinBase64_non_string_non_array.jsonnet
… testdata/builtinBase64_string_high_codepoint.jsonnet