```
benchmark                                 old ns/op       new ns/op     delta
Benchmark_Builtin_base64-16               10805730974     23158636      -99.79%
Benchmark_Builtin_base64_bytearray-16     8682808704      123360964     -98.58%
```

### Notes

~This PR contains commits from another pending PR. I'll rebased once #343 is merged.~