@ghostsquad 2019-12-15
resolves #341