@ghostsquad 2020-01-26
@sbarzowski this has been rebased and is ready for review/merge